+++ astropy/time/core.py
@@ -1503,10 +1503,12 @@ class TimeDelta(Time):
 
     Parameters
     ----------
-    val : numpy ndarray, list, str, number, or `~astropy.time.TimeDelta` object
-        Data to initialize table.
-    val2 : numpy ndarray, list, str, or number; optional
-        Data to initialize table.
+    val : sequence, ndarray, number, `~astropy.units.Quantity` or `~astropy.time.TimeDelta` object
+        Value(s) to initialize the time difference(s). Any quantities will
+        be converted appropriately (with care taken to avoid rounding
+        errors for regular time units).
+    val2 : sequence, ndarray, number, or `~astropy.units.Quantity`; optional
+        Additional values, as needed to preserve precision.
     format : str, optional
         Format of input value(s)
     scale : str, optional
@@ -1531,14 +1533,6 @@ class TimeDelta(Time):
                 self._set_scale(scale)
         else:
             if format is None:
-                try:
-                    val = val.to(u.day)
-                    if val2 is not None:
-                        val2 = val2.to(u.day)
-                except Exception:
-                    raise ValueError('Only Quantities with Time units can '
-                                     'be used to initiate {0} instances .'
-                                     .format(self.__class__.__name__))
                 format = 'jd'
 
             self._init_from_vals(val, val2, format, scale, copy)