+++ astropy/time/formats.py
@@ -15,7 +15,7 @@ from .. import units as u
 from .. import _erfa as erfa
 from ..extern import six
 from ..extern.six.moves import zip
-from .utils import day_frac, two_sum
+from .utils import day_frac, quantity_day_frac, two_sum, two_product
 
 
 __all__ = ['TimeFormat', 'TimeJD', 'TimeMJD', 'TimeFromEpoch', 'TimeUnix',
@@ -154,11 +154,31 @@ class TimeFormat(object):
                             .format(self.name))
 
         if getattr(val1, 'unit', None) is not None:
-            # Possibly scaled unit any quantity-likes should be converted to
-            _unit = u.CompositeUnit(getattr(self, 'unit', 1.), [u.day], [1])
-            val1 = u.Quantity(val1, copy=False).to_value(_unit)
+            # Convert any quantity-likes to days first, attempting to be
+            # careful with the conversion, so that, e.g., large numbers of
+            # seconds get converted without loosing precision because
+            # 1/86400 is not exactly representable as a float.
+            val1 = u.Quantity(val1, copy=False)
             if val2 is not None:
-                val2 = u.Quantity(val2, copy=False).to_value(_unit)
+                val2 = u.Quantity(val2, copy=False)
+
+            try:
+                val1, val2 = quantity_day_frac(val1, val2)
+            except u.UnitsError:
+                raise u.UnitConversionError(
+                    "only quantities with time units can be "
+                    "used to instantiate Time instances.")
+            # We now have days, but the format may expect another unit.
+            # On purpose, multiply with 1./day_unit because typically it is
+            # 1./erfa.DAYSEC, and inverting it recovers the integer.
+            # (This conversion will get undone in format's set_jds, hence
+            # there may be room for optimizing this.)
+            factor = 1. / getattr(self, 'unit', 1.)
+            if factor != 1.:
+                val1, carry = two_product(val1, factor)
+                carry += val2 * factor
+                val1, val2 = two_sum(val1, carry)
+
         elif getattr(val2, 'unit', None) is not None:
             raise TypeError('Cannot mix float and Quantity inputs')